Allow quoted filter keyword arguments.
In d706db3bd70ac support of keyworded arguments
was added and made the workaround obsolete.

The workaround implied that you convert the given arguments
arry into a hash using Array#to_hash. However this
is not necessary anymore. Liquid supports this
notation

    {{ variable | filter: key1: "value", key2: "value2" }}

The filter would look like this:

    module Filter
      def filter(options = {})
        puts options.inspect #=> { "key1" => "value", "key2"=> ..
      end
    end

But this is not suffcient for Bridge templates since the workaround
also supported quotes keywords like this

     {{ leads.check.form_field | leads_options: "tabindex": 1 }}

And this notation is widely used. This commit ensures backwards
compatility to our templates.